chore: reorganize the codebase
