+++ packages/client/src/stats/rtc/index.ts
@@ -1,0 +1,5 @@
+// Based on the original code from:
+// https://github.com/fippo/rtcstats
+
+export * from './pc';
+export * from './TraceBuffer';
