+++ packages/client/src/stats/rtc/utils.ts
@@ -57,15 +57,3 @@ export const deltaCompression = (
   newStats.timestamp = timestamp;
   return newStats;
 };
-
-export const dumpStream = (stream: MediaStream) => ({
-  id: stream.id,
-  tracks: stream.getTracks().map((track) => ({
-    id: track.id, // unique identifier (GUID) for the track
-    kind: track.kind, // `audio` or `video`
-    label: track.label, // identified the track source
-    enabled: track.enabled, // application can control it
-    muted: track.muted, // application cannot control it (read-only)
-    readyState: track.readyState, // `live` or `ended`
-  })),
-});
